+++ source/extensions/isaacsim.examples.interactive/isaacsim/examples/interactive/kaya_gamepad/kaya_gamepad_extension.py
@@ -16,8 +16,8 @@
 import os
 
 import omni.ext
+from isaacsim.base_sample.base_sample_extension_experimental import BaseSampleUITemplate
 from isaacsim.examples.browser import get_instance as get_browser_instance
-from isaacsim.examples.interactive.base_sample import BaseSampleUITemplate
 from isaacsim.examples.interactive.kaya_gamepad import KayaGamepad
 
 
@@ -44,14 +44,9 @@ class KayaGamepadExtension(omni.ext.IExt):
         # register the example with examples browser
         get_browser_instance().register_example(
             name=self.example_name,
-            execute_entrypoint=ui_handle.build_window,
             ui_hook=ui_handle.build_ui,
             category=self.category,
         )
 
-        return
-
     def on_shutdown(self):
         get_browser_instance().deregister_example(name=self.example_name, category=self.category)
-
-        return